env_requirements.txt
absl-py==2.1.0
asdf==3.2.0
asdf_standard==1.1.1
asdf_transform_schemas==0.5.0
astunparse==1.6.3
attrs==23.2.0
awkward==2.6.5
awkward-cpp==34
blinker==1.7.0
boost_histogram==1.4.1
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
colored==2.2.4
colorlog==6.8.2
contourpy==1.2.1
cramjam==2.8.3
cycler==0.12.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
dm-tree==0.1.8
dotmap==1.3.30
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.53.0
frozendict==2.4.4
fsspec==2024.5.0
gast==0.5.4
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.0
h5py==3.11.0
hep-ml==0.7.2
hepstats==0.8.1
hist==2.7.3
histoprint==2.4.0
idna==3.7
iminuit==2.26.0
importlib_metadata==7.0.2
importlib_resources==6.4.0
itsdangerous==2.1.2
jacobi==0.9.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
meson==1.3.2
ml-dtypes==0.3.2
mplhep==0.3.48
mplhep-data==0.0.3
namex==0.0.8
ninja==1.11.1.1
numdifftools==0.9.41
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
ordered-set==4.1.0
packaging==24.0
pandas==2.2.1
pillow==10.3.0
pkgconfig==1.5.5
protobuf==4.25.3
pyarrow==15.0.1
pydantic==1.10.15
Pygments==2.18.0
PyMySQL==1.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
six==1.16.0
SQLAlchemy==2.0.28
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-probability==0.24.0
termcolor==2.4.0
texttable==1.7.0
tf_keras==2.16.0
Theano==1.0.5
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.10.0
tzdata==2024.1
uhi==0.4.0
uproot==5.3.7
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.16.0
xxhash==3.4.1
zfit==0.20.3
zfit-interface==0.0.3
zipp==3.17.0
